fix: 盘存-验收问题

+++ app/src/main/java/com/sunmi/uhf/fragment/readwrite/ReadWriteViewModel.kt
@@ -20,7 +20,7 @@ class ReadWriteViewModel : BaseViewModel() {
     val totalLabelNum = MutableLiveData<Int>(0)
 
     /* 读取速度 */
-    val speed = MutableLiveData<Float>(0f)
+    val speed = MutableLiveData<Int>(0)
 
     /* 是否开始读取 */
     val start = MutableLiveData<Boolean>(false)
